use metered src/sink

+++ pkg/stream/processor/sink.go
@@ -3,5 +3,5 @@ package processor
 import "sharedlog-stream/pkg/stream/processor/commtypes"
 
 type Sink interface {
-	Sink(msg commtypes.Message, parNum uint32) error
+	Sink(msg commtypes.Message, parNum uint8) error
 }
